refactor: change tests to use node's test runner
Renaming the directory since the underscores seem
excessive.

Refs: #19
