+++ app/tests/cli.test.js
@@ -1,0 +1,12 @@
+import { describe, it } from "node:test"
+import assert from "node:assert"
+
+describe('cli', () => {
+  // TODO: Add tests after Jest supports ES modules.
+  // @see https://github.com/facebook/jest/pull/10976
+  it('should return true', () => {
+    const placeholder = true;
+
+    assert.strictEqual(placeholder, true);
+  });
+});
